Разработал компонент отображения подробных результатов соревнования по стадиям. Добавил методы для вытягивания информации по стадиям и по участникам соревнования в контроллере Competition и в соответствующем сервисе.

+++ Models/Circle.cs
@@ -7,7 +7,7 @@ namespace circle_competitions_monitoring.Models {
         public string circle_name { get; set; }
         public int sportsman { get; set; }
         public TimeSpan time_of_finish { get; set; }
-        public float points { get; set; }
+        public double points { get; set; }
         public int place { get; set; }
     }
 }
